Refactor ProjectGridItem component for improved rendering and structure
This commit refactors the `ProjectGridItem.vue` component to enhance its rendering logic and maintainability. The component now uses `v-if` and `v-else` for consistent rendering of the `NuxtLink` and fallback `div`, simplifying the structure. Additionally, the mobile-only button is wrapped in a `ClientOnly` component to ensure proper client-side rendering. These changes aim to improve the overall clarity and functionality of the component while adhering to best practices in Vue.js development.